OpenCV Limitation
Wanted to change the icon of the camera window to make it look better but this cannot be done with just OpenCV and Python. https://stackoverflow.com/questions/60259324/is-there-any-method-to-change-the-icon-for-the-imshow-window. For now I will ignore this as it is not necessary as it's not part of the success criteria, and it can always be added at a later date.