Flip camera
Was trying to find ranges for rgb values on the cube, but i was finding it challenging as the camera was the different way round than most of us are used to. To fix this i flipped the frame, only needs one line of code. I learnt this code from the "Saving a video" section in this part of the opencv docs
https://docs.opencv.org/master/dd/d43/tutorial_py_video_display.html