Output colours on face
Now when space is pressed, the colours detected will be printed. I had to switch the nested for loops around as it meant the colours were order different to how i wanted (down columns instead of across rows). I then returned the string after the scanFace() function had finished. Also i made cv.waitkey its own variable because i notice if i just make another if statement using if cv.waitkey == ..... then it doesnt work so well. I made it when the user pressed space (ascii code 32), the face is printed in the form of characters representing the first letter of that colour